refactor: split typescript language
BREAKING CHANGE: typescript template was moved to `typescript/2.0-rx/index.ts`
