+++ src/language/typescript/2.0-rx/data/serialized-path-parameter.ts
@@ -1,0 +1,22 @@
+import { SerializedDependency } from './serialized-dependency';
+import { SerializedParameter } from './serialized-parameter';
+
+export interface SerializedPathParameter extends SerializedParameter {
+	readonly name: string;
+}
+
+export const serializedPathParameter = (
+	name: string,
+	type: string,
+	io: string,
+	isRequired: boolean,
+	dependencies: SerializedDependency[],
+	refs: string[],
+): SerializedPathParameter => ({
+	name,
+	type,
+	io,
+	isRequired,
+	dependencies,
+	refs,
+});
