+++ src/language/typescript/2.0-rx/index.ts
@@ -1,0 +1,1 @@
+export { serializeSwaggerObject as serialize } from './serializers/swagger-object';
