+++ src/language/typescript/2.0-rx/serializers/paths-object.ts
@@ -1,0 +1,16 @@
+import { PathsObject } from '../../../../schema/2.0/paths-object';
+import { Option } from 'fp-ts/lib/Option';
+import { ParametersDefinitionsObject } from '../../../../schema/2.0/parameters-definitions-object';
+import { directory, Directory } from '../../../../fs';
+import { CONTROLLERS_DIRECTORY, ROOT_DIRECTORY } from '../utils';
+import { serializeDictionary } from '../../../../utils/types';
+import { groupPathsByTag } from '../../../../utils';
+import { serializePathGroup } from './path-item-object';
+
+export const serializePaths = (paths: PathsObject, parameters: Option<ParametersDefinitionsObject>): Directory =>
+	directory(
+		CONTROLLERS_DIRECTORY,
+		serializeDictionary(groupPathsByTag(paths, parameters), (name, group) =>
+			serializePathGroup(name, group, `${ROOT_DIRECTORY}/${CONTROLLERS_DIRECTORY}`),
+		),
+	);
