+++ src/utils/monoid.ts
@@ -1,0 +1,3 @@
+import { getMonoid as getArrayMonoid } from 'fp-ts/lib/Array';
+
+export const monoidStrings = getArrayMonoid<string>();
