+++ src/utils/string.ts
@@ -1,0 +1,2 @@
+export const unless = (condition: boolean, a: string): string => (condition ? '' : a);
+export const when = (condition: boolean, a: string): string => (condition ? a : '');
