+++ src/utils/types.ts
@@ -1,1 +1,6 @@
 export interface Dictionary<A> extends Readonly<Record<string, A>> {}
+
+export const serializeDictionary = <A, B>(
+	dictionary: Dictionary<A>,
+	serializeValue: (name: string, value: A) => B,
+): B[] => Object.keys(dictionary).map(name => serializeValue(name, dictionary[name]));
